MIGRATIONS - fixed typo and
Fixed typo and doing migrations when it comes to modifications using the proper
method. After all it is good to get in the habit of creating migrations when
changes happen to easier track why. As far as why the is_partner column does
not fit this update when compared with the original version of this application