MODEL_MIGRATIONS - created a few more migrations
Created a few more migration files, for user roles and user permissions.
Removing the role_id and permission_id from the users table. Allowing once
again a user to have many roles and many permissions. And better defined the
relationships between the tables in the user, property, and booking tables