ROLES_PERMISSIONS - permissions is a new concept
Permissions is a new concept that was never built into the original application.
Most of the roles were carried over however I feel like it is going to need
more refinement, considering adding permissions. The initial application
allowed a user to have many roles, however it never went pass the one to one
relationship but was built in, in the event it was